fix up login card a little bit, its still a pretty much just thrown together
