+++ apps/designer/app/auth/login.tsx
@@ -11,7 +11,7 @@ import {
 } from "@webstudio-is/design-system";
 import interStyles from "~/shared/font-faces/inter.css";
 import { GithubIcon, CommitIcon, GoogleIcon } from "@webstudio-is/icons";
-import { LoginButton } from "./components/login-button";
+import { LoginButton } from "./login-button";
 import loginStyles from "./login.css";
 import { authPath } from "~/shared/router-utils";
 
@@ -46,7 +46,7 @@ export const Login = ({ errorMessage }: { errorMessage: string }) => {
       <Card size={2}>
         <Flex direction="column" align="center" gap="3">
           <Heading size="2">Login</Heading>
-          <Flex direction="column" gap="4" css={{ width: 300 }}>
+          <Flex direction="column" gap="4">
             {errorMessage.length ? (
               <Text align="center" color="error">
                 {errorMessage}
@@ -54,23 +54,19 @@ export const Login = ({ errorMessage }: { errorMessage: string }) => {
             ) : null}
             <Flex gap="2" direction="column">
               <Form action={authPath({ provider: "github" })} method="post">
-                <LoginButton enabled={data.isGithubEnabled}>
-                  <Flex gap="2" align="center">
-                    <Flex css={{ width: 20, height: 20 }}>
-                      <GithubIcon />
-                    </Flex>
-                    Login with GitHub
-                  </Flex>
+                <LoginButton
+                  disabled={data.isGithubEnabled === false}
+                  icon={<GithubIcon />}
+                >
+                  Login with GitHub
                 </LoginButton>
               </Form>
               <Form action={authPath({ provider: "google" })} method="post">
-                <LoginButton enabled={data.isGoogleEnabled}>
-                  <Flex gap="2" align="center">
-                    <Flex css={{ width: 20, height: 20 }}>
-                      <GoogleIcon />
-                    </Flex>
-                    Login with Google
-                  </Flex>
+                <LoginButton
+                  disabled={data.isGoogleEnabled === false}
+                  icon={<GoogleIcon />}
+                >
+                  Login with Google
                 </LoginButton>
               </Form>
               {data.isDevLogin &&
@@ -97,16 +93,12 @@ export const Login = ({ errorMessage }: { errorMessage: string }) => {
                   </Flex>
                 ) : (
                   <LoginButton
-                    enabled={data.isDevLogin}
+                    disabled={data.isDevLogin === false}
                     isDevLogin
                     onClick={() => openDevLogin(true)}
+                    icon={<CommitIcon />}
                   >
-                    <Flex gap="2" align="center">
-                      <Flex css={{ width: 20, height: 20 }}>
-                        <CommitIcon />
-                      </Flex>
-                      Dev Login
-                    </Flex>
+                    Dev Login
                   </LoginButton>
                 ))}
             </Flex>
